separate Header and side menu
